agents/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langgraph==0.0.55
langchain==0.1.16
langchain-openai==0.1.3
paramiko==3.4.0
cryptography==42.0.0
aiosqlite==0.20.0
pydantic==2.7.0
python-dotenv==1.0.1
